Moved doxygen_pragma out of public headers
