+++ sdk/core/azure-core-amqp/src/private/doxygen_pragma.hpp
